sketch for using chained Q-objects
This whole attempt is a failure because I was too unstructured since a
certain point.

I also need to concentrate on which kind of interface I intend to
offer and proceed more carefully.

I've documented the lessons learned for myself and will try to
identify the point where the problems started and proceed from there.

I think the initial and some of the later refactorings were OK, so I
probably have to go on another spike from there onwards to see what
the next steps are.

The general approach in this commit looks promising, but I've
identified the exact logic I need to replicate which will need a
slightly different interface.

(The global search query goes into a OR-disjunction for all columns, the
column searches need to be combined with AND on top of that. This
means I need to track the global search separately. I'm not sure yet
how this could be made compatible with the django-filter approach.)